refactor: move StaleNodeError to Sdk
