+++ packages/rs-sdk/src/error.rs
@@ -67,6 +67,10 @@ pub enum Error {
     /// Operation cancelled - cancel token was triggered, timeout, etc.
     #[error("Operation cancelled: {0}")]
     Cancelled(String),
+
+    /// Remote node is stale; try another server
+    #[error(transparent)]
+    StaleNode(#[from] StaleNodeError),
 }
 
 impl<T: Debug + Mockable> From<DapiClientError<T>> for Error {
@@ -85,10 +89,37 @@ impl CanRetry for Error {
     fn can_retry(&self) -> bool {
         matches!(
             self,
-            Error::Proof(drive_proof_verifier::Error::StaleNode(..))
+            Error::StaleNode(..)
                 | Error::DapiClientError(_)
                 | Error::CoreClientError(_)
                 | Error::TimeoutReached(_, _)
         )
     }
+}
+
+/// Server returned stale metadata
+#[derive(Debug, thiserror::Error)]
+pub enum StaleNodeError {
+    /// Server returned metadata with outdated height
+    #[error("received height is outdated: expected {expected_height}, received {received_height}, tolerance {tolerance_blocks}; try another server")]
+    Height {
+        /// Expected height - last block height seen by the Sdk
+        expected_height: u64,
+        /// Block height received from the server
+        received_height: u64,
+        /// Tolerance - how many blocks can be behind the expected height
+        tolerance_blocks: u64,
+    },
+    /// Server returned metadata with time outside of the tolerance
+    #[error(
+        "received invalid time: expected {expected_timestamp_ms}ms, received {received_timestamp_ms} ms, tolerance {tolerance_ms} ms; try another server"
+    )]
+    Time {
+        /// Expected time in milliseconds - is local time when the message was received
+        expected_timestamp_ms: u64,
+        /// Time received from the server in the message, in milliseconds
+        received_timestamp_ms: u64,
+        /// Tolerance in milliseconds
+        tolerance_ms: u64,
+    },
 }
